+++ app/javascript/vue/tasks/taxon_names/filter/components/list.vue
@@ -3,19 +3,11 @@
     <table class="full_width">
       <thead>
         <tr>
-          <th
-            :class="classSort('name')"
-            @click="sortTable('name')">Name</th>
-          <th
-            :class="classSort('verbatim_author')"
-            @click="sortTable('verbatim_author')">Author and year</th>
-          <th
-            :class="classSort('original_combination')"
-            @click="sortTable('original_combination')">Original combination</th>
+          <th @click="sortTable('name')">Name</th>
+          <th @click="sortTable('cached_author_year')">Author and year</th>
+          <th @click="sortTable('original_combination')">Original combination</th>
           <th>Valid?</th>
-          <th
-            :class="classSort('rank')"
-            @click="sortTable('rank')">Rank</th>
+          <th @click="sortTable('rank')">Rank</th>
           <th>Parent</th>
           <th>Options</th>
         </tr>
@@ -74,37 +66,16 @@ export default {
   },
   data() {
     return {
-      sortColumns: {
-        name: undefined,
-        verbatim_author: undefined,
-        year_of_publication: undefined,
-        rank: undefined,
-        original_combination: undefined
-      }
+      ascending: false
     }
   },
   methods: {
-    sortTable(sortProperty) {
-      let that = this
-      function compare(a,b) {
-        if (a[sortProperty] < b[sortProperty])
-          return (that.sortColumns[sortProperty] ? -1 : 1)
-        if (a[sortProperty] > b[sortProperty])
-          return (that.sortColumns[sortProperty] ? 1 : -1);
-        return 0
-      }
-      if(this.sortColumns[sortProperty] == undefined) {
-        this.sortColumns[sortProperty] = true
-      }
-      else {
-        this.sortColumns[sortProperty] = !this.sortColumns[sortProperty]
-      }
-      this.list.sort(compare)
-    },
-    classSort(value) {
-      if(this.sortColumns[value] == true) { return 'headerSortDown' }
-      if(this.sortColumns[value] == false) { return 'headerSortUp' }
-      return ''
+    sortTable (sortProperty) {
+      this.list.sort((a, b) => this.ascending ?
+        a[sortProperty]?.localeCompare(b[sortProperty], 'en', {'sensitivity': 'base'}) :
+        b[sortProperty]?.localeCompare(a[sortProperty], 'en', {'sensitivity': 'base'})
+      )
+      this.ascending = !this.ascending
     }
   }
 }